# swagger.yaml
swagger: "2.0"
info:
  title: Space Mission Control API
  version: 1.0.1
  description: |
    API for managing space missions, control users, astronauts, and launch vehicles.
    All error strings are required to be clear descriptive message. The descriptive message won't change the marks you receive when being automarked as long as you make an attempt for each one to be custom, but they will affect your code quality mark.
schemes:
  - http
...
components:
  prim:
    ControlUserSessionId:
      type: string
      description: The sessionID representing a logged in controlUser
      example: "abcd1234"
    Error:
      type: string
      example: "Meaningful error message"
    Email:
      type: string
      example: james.kirk@starfleet.gov.au
...
    MissionId:
      type: number
      example: 1
    AstronautId:
      type: number
      example: 1
...
    AstronautDesignation:
      type: string
      description: A concatenated string of rank, nameFirst, space and then nameLast
      example: Captain James Kirk
    Timestamp:
      type: number
      example: 1739912980
      description: In seconds. You can find more information that here https://en.wikipedia.org/wiki/Unix_time
...
  group:
...
    SmallMissions:
      type: array
      items:
        type: object
        properties:
          missionId:
            $ref: '#/components/prim/MissionId'
          name:
            type: string
            example: My Space Mission name
...
    MissionIter3:
      type: object
      properties:
        missionId:
          $ref: '#/components/prim/MissionId'
        name:
          type: string
          example: This is the name of the space mission
        timeCreated:
          $ref: '#/components/prim/Timestamp'
        timeLastEdited:
          $ref: '#/components/prim/Timestamp'
        description:
          type: string
          example: A description of the aims of this space mission
        target:
          type: string
          example: The target destination of this space mission
        assignedAstronauts:
          type: array
          items:
            type: object
            properties:
              astronautId:
                type: number
                example: 1
              designation:
                $ref: '#/components/prim/AstronautDesignation'
...
  path:
    MissionId:
      in: path
      name: missionid
      required: true
      schema:
        $ref: '#/components/prim/MissionId'
    AstronautId:
      in: path
      name: astronautid
      required: true
      schema:
        $ref: '#/components/prim/AstronautId'
...
  header:
    ControlUserSessionId:
      in: header
      name: controlUserSessionId
      required: true
      schema:
        $ref: '#/components/prim/ControlUserSessionId'
  body:
...
    MissionCreate:      
      in: body
      name: body
      required: true
      schema:
        type: object
        properties:
          name:
            type: string
          description:
            type: string
            example: A description of this space mission
          target:
            type: string
            example: The target destination of this space mission
    MissionNameUpdate:      
      in: body
      name: body
      required: true
      schema:
        type: object
        properties:
          name:
            type: string
    MissionDescriptionUpdate:      
      in: body
      name: body
      required: true
      schema:
        type: object
        properties:
          description:
            type: string
            example: A description of this space mission
    MissionTargetUpdate:      
      in: body
      name: body
      required: true
      schema:
        type: object
        properties:
          target:
            type: string
            example: The target destination of this space mission
    MissionTransfer:
      in: body
      name: body
      required: true
      schema:
        type: object
        properties:
          controlUserEmail:
            $ref: '#/components/prim/Email'
...
  return:
...
    Error:
      properties:
        error:
          $ref: '#/components/prim/Error'
    Empty:
      type: object
...
    MissionList:
      type: object
      properties:
        missions:
          $ref: '#/components/group/SmallMissions'
    MissionNew:
      type: object
      properties:
        missionId:
          $ref: '#/components/prim/MissionId'
...
    MissionSingleIter3:
      $ref: '#/components/group/MissionIter3' 
...
paths:
...
  /v1/admin/mission/list:
    get:
      summary: List all missions for the current control user.
      tags:
        # - "By Topic: Mission Management"
        - "Iteration 2 (Using Iteration 1)"
      parameters:
        - $ref: '#/components/header/ControlUserSessionId'
      responses:
        200:
          description: OK
          schema:
            $ref: '#/components/return/MissionList'
        401:
          description: ControlUserSessionId is empty or invalid (does not refer to valid logged in user session)
          schema:
            $ref: '#/components/return/Error'
  /v1/admin/mission:
    post:
      summary: Create a new space mission
      description: Given the basic details for a space mission, create one for the logged in control user.
      tags:
        # - "By Topic: Mission Management"
        - "Iteration 2 (Using Iteration 1)"
      parameters:
        - $ref: '#/components/header/ControlUserSessionId'
        - $ref: '#/components/body/MissionCreate'
      responses:
        200:
          description: OK
          schema:
            $ref: '#/components/return/MissionNew'
        401:
          description: ControlUserSessionId is empty or invalid (does not refer to valid logged in user session)
          schema:
            $ref: '#/components/return/Error'
        400:
          description: |
            If any of the following are true:
            - Name contains invalid characters. Valid characters are alphanumeric and spaces
            - Name is either less than 3 characters long or more than 30 characters long
            - Name is already used by the current logged in user for another space mission
            - Description is more than 400 characters in length (note: empty strings are OK)
            - Target is more than 100 characters in length (note: empty strings are OK)
          schema:
            $ref: '#/components/return/Error'
  /v1/admin/mission/{missionid}:
    delete:
      summary: Remove a mission
      description: |
       - Given a particular space mission, delete it.
      tags:
        # - "By Topic: Mission Management"
        - "Iteration 2 (Using Iteration 1)"
      parameters:
        - $ref: '#/components/header/ControlUserSessionId'
        - $ref: '#/components/path/MissionId'
      responses:
        200:
          description: OK
          schema:
            $ref: '#/components/return/Empty'
        401:
          description: ControlUserSessionId is empty or invalid (does not refer to valid logged in user session)
          schema:
            $ref: '#/components/return/Error'
        403:
          description: Valid controlUserSessionId is provided, but control user is not an owner of this space mission or the space mission doesn't exist
          schema:
            $ref: '#/components/return/Error'
        400:
          description: |
            If any of the following are true:
            - Astronauts have been assigned to this mission.
          schema:
            $ref: '#/components/return/Error'

    get:
      summary: Get info about current space mission
      description: Get all of the relevant information about the current space mission including attached astronauts.
      tags:
        # - "By Topic: Mission Management"
        - "Iteration 2 (Using a modified Iteration 1)"
      parameters:
        - $ref: '#/components/header/ControlUserSessionId'
        - $ref: '#/components/path/MissionId'
      responses:
        200:
          description: OK
          schema:
            $ref: '#/components/return/MissionSingleIter3'
        401:
          description: ControlUserSessionId is empty or invalid (does not refer to valid logged in user session)
          schema:
            $ref: '#/components/return/Error'
        403:
          description: Valid controlUserSessionId is provided, but the control user is not an owner of this mission or the specified missionId does not exist
          schema:
            $ref: '#/components/return/Error'
  /v1/admin/mission/{missionid}/name:
    put:
      summary: Update the name of the specified space mission
      description: Given a new name for a space mission, update the name for the specified space mission
      tags:
        # - "By Topic: Access Management"
        - "Iteration 2 (Using Iteration 1)"
      parameters:
        - $ref: '#/components/header/ControlUserSessionId'
        - $ref: '#/components/path/MissionId'
        - $ref: '#/components/body/MissionNameUpdate'
      responses:
        200:
          description: OK
          schema:
            $ref: '#/components/return/Empty'
        401:
          description: ControlUserSessionId is empty or invalid (does not refer to valid logged in user session)
          schema:
            $ref: '#/components/return/Error'
        403:
          description: Valid controlUserSessionId is provided, but the control user is not an owner of this mission or the specified missionId does not exist
          schema:
            $ref: '#/components/return/Error'
        400:
          description: |
            If any of the following are true:
            - Name contains invalid characters. Valid characters are alphanumeric and spaces
            - Name is either less than 3 characters long or more than 30 characters long
            - Name is already used by the current logged in user for another quiz
          schema:
            $ref: '#/components/return/Error'
  /v1/admin/mission/{missionid}/description:
    put:
      summary: Update the description of the specified space mission
      description: Given a new description for a space mission, update the description for the specified space mission
      tags:
        # - "By Topic: Access Management"
        - "Iteration 2 (Using Iteration 1)"
      parameters:
        - $ref: '#/components/header/ControlUserSessionId'
        - $ref: '#/components/path/MissionId'
        - $ref: '#/components/body/MissionDescriptionUpdate'
      responses:
        200:
          description: OK
          schema:
            $ref: '#/components/return/Empty'
        401:
          description: ControlUserSessionId is empty or invalid (does not refer to valid logged in user session)
          schema:
            $ref: '#/components/return/Error'
        403:
          description: Valid controlUserSessionId is provided, but the control user is not an owner of this mission or the specified missionId does not exist
          schema:
            $ref: '#/components/return/Error'
        400:
          description: |
            If any of the following are true:
            - Description is more than 400 characters in length (note: empty strings are OK)
          schema:
            $ref: '#/components/return/Error'
  /v1/admin/mission/{missionid}/target:
    put:
      summary: Update the target of the specified space mission
      description: Given a new target for a space mission, update the target for the specified space mission
      tags:
        # - "By Topic: Access Management"
        - "Iteration 2 (Using Iteration 1)"
      parameters:
        - $ref: '#/components/header/ControlUserSessionId'
        - $ref: '#/components/path/MissionId'
        - $ref: '#/components/body/MissionTargetUpdate'
      responses:
        200:
          description: OK
          schema:
            $ref: '#/components/return/Empty'
        401:
          description: ControlUserSessionId is empty or invalid (does not refer to valid logged in user session)
          schema:
            $ref: '#/components/return/Error'
        403:
          description: Valid controlUserSessionId is provided, but the control user is not an owner of this mission or the specified missionId does not exist
          schema:
            $ref: '#/components/return/Error'
        400:
          description: |
            If any of the following are true:
            - Target is more than 100 characters in length (note: empty strings are OK)
          schema:
            $ref: '#/components/return/Error'

  # /v2/admin/mission/{missionid}/transfer:
  #   post:
  #     summary: Transfer this mission to another control user
  #     description: Given the email of another control user, transfer ownership of this mission to that new control user.
  #     tags:
  #       - "Iteration 3 (Using a modified Iteration 2)"
  #     parameters:
  #       - $ref: '#/components/header/ControlUserSessionId'
  #       - $ref: '#/components/path/MissionId'
  #       - $ref: '#/components/body/MissionTransfer'
  #     responses:
  #       200:
  #         description: OK
  #         schema:
  #           $ref: '#/components/return/Empty'
  #       401:
  #         description: ControlUserSessionId is empty or invalid (does not refer to valid logged in user session)
  #         schema:
  #           $ref: '#/components/return/Error'
  #       403:
  #         description: Valid ControlUserSessionId is provided, but control user is not an owner of this space mission or space mission doesn't exist
  #         schema:
  #           $ref: '#/components/return/Error'
  #       400:
  #         description: |
  #           If any of the following are true:
  #           - userEmail is not a real control user
  #           - userEmail is the current logged in control user
  #           - missionId refers to a space mission that has a name that is already used by the target user
  #         schema:
  #           $ref: '#/components/return/Error'
...
  /v1/admin/mission/{missionid}/assign/{astronautid}:
    post:
      summary: Assign an astronaut to a space mission.
      tags:
        # - "By Topic: Mission Management"
        - "Iteration 2 (New)"
      parameters:
        - $ref: '#/components/header/ControlUserSessionId'
        - $ref: '#/components/path/AstronautId'
        - $ref: '#/components/path/MissionId'
      responses:
        200:
          description: OK
          schema:
            $ref: '#/components/return/Empty'
        401:
          description: ControlUserSessionId is empty or invalid (does not refer to valid logged in user session)
          schema:
            $ref: '#/components/return/Error'
        403:
          description: Valid controlUserSessionId is provided, but the control user is not an owner of this mission or the specified missionId does not exist
          schema:
            $ref: '#/components/return/Error'
        400:
          description: |
            If any of the following are true:
            - astronautid is invalid
            - The astronaut is already assigned to another mission
          schema:
            $ref: '#/components/return/Error'
            
    delete:
      summary: Un-assign an astronaut from a space mission.
      tags:
        # - "By Topic: Mission Management"
        - Iteration 2 (New)
      parameters:
        - $ref: '#/components/header/ControlUserSessionId'
        - $ref: '#/components/path/AstronautId'
        - $ref: '#/components/path/MissionId'
      responses:
        200:
          description: OK
          schema:
            $ref: '#/components/return/Empty'
        401:
          description: ControlUserSessionId is empty or invalid (does not refer to valid logged in user session)
          schema:
            $ref: '#/components/return/Error'
        403:
          description: Valid controlUserSessionId is provided, but the control user is not an owner of this mission or the specified missionId does not exist
          schema:
            $ref: '#/components/return/Error'
        400:
          description: |
            If any of the following are true:
            - astroanutid is invalid
            - The astronaut not assigned to this space mission.
          schema:
            $ref: '#/components/return/Error'
  /v2/admin/mission/{missionid}/assign/{astronautid}:
    delete:
      summary: Un-assign an astronaut from a space mission.
      tags:
        # - "By Topic: Mission Management"
        - Iteration 3 (Using a modified Iteration 2)
      parameters:
        - $ref: '#/components/header/ControlUserSessionId'
        - $ref: '#/components/path/AstronautId'
        - $ref: '#/components/path/MissionId'
      responses:
        200:
          description: OK
          schema:
            $ref: '#/components/return/Empty'
        401:
          description: ControlUserSessionId is empty or invalid (does not refer to valid logged in user session)
          schema:
            $ref: '#/components/return/Error'
        403:
          description: Valid controlUserSessionId is provided, but the control user is not an owner of this mission or the specified missionId does not exist
          schema:
            $ref: '#/components/return/Error'
        400:
          description: |
            If any of the following are true:
            - astroanutid is invalid
            - The astronaut not assigned to this space mission.
            - The astronaut is allocated to a launch (NEW ERROR)
          schema:
            $ref: '#/components/return/Error'
...